% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polimerizadorParalelo.R
\name{PolyMaS}
\alias{PolyMaS}
\title{This package performs the computational polymerization of linear molecules from their structural repetitive units by using Simplified Molecular Input Line Entry Specification (SMILES).}
\usage{
PolyMaS(SMILES_SRU, PD)
}
\arguments{
\item{SMILES_SRU}{The Structural Repetitive Unit expresed by SMILES code. It must contain two asterisks (*), one indicating the head and the other indicating the tail.}

\item{PD}{Polymerization Degree.}
}
\value{
The SMILES representation of the resulting polymer with the desired polymerization degree.
}
\description{
This package performs the computational polymerization of linear molecules from their structural repetitive units by using Simplified Molecular Input Line Entry Specification (SMILES).
}
\examples{
PolyMaS("*C(C*)c1ccccc1",19)
}
